requirements.txt
flask==2.3.2
supabase==2.3.1
python-dotenv==1.0.0
scikit-learn==1.3.0
nltk==3.8.1
numpy==1.24.3
pandas==2.0.3
python-Levenshtein==0.21.1
gunicorn==20.1.0
joblib==1.3.2
spacy>=3.4.0
requests==2.31.0
werkzeug==2.3.7
flask-cors==4.0.0
psutil==5.9.5
